Cleanup
Removes RecordingPaintedItem
Removes INetQPainter (functionality moved to NetQPainter)
