+++ QmlNet/controls/NetQPainter.h
@@ -3,76 +3,77 @@
 
 #include <QPainter>
 #include <map>
-#include "INetQPainter.h"
 
-class NetQPainter : public virtual INetQPainter
+class NetQPainter
 {
 public:
     NetQPainter(QPainter* qpainter);
-    ~NetQPainter() override;
+    ~NetQPainter();
 
-    void setPen(int colorId) override;
-    void resetPen() override;
-    void setBrush(int colorId) override;
-    void resetBrush() override;
-    void setFont(int fontFamilyId, bool isBold, bool isItalic, bool isUnderline, int pxSize) override;
-    void setFontFamily(int fontFamilyId) override;
-    void setFontBold(bool isBold) override;
-    void setFontItalic(bool isItalic) override;
-    void setFontUnderline(bool isUnderline) override;
-    void setFontSize(int pxSize) override;
-    void drawText(int x, int y, QString text) override;
-    void drawText(int x, int y, int width, int height, int flags, QString text) override;
-    void drawRect(int x, int y, int width, int height) override;
-    void fillRect(int x, int y, int width, int height, int colorId) override;
-    void fillRect(int x, int y, int width, int height) override;
-    void drawArc(int x, int y, int width, int height, int startAngle, int spanAngle) override;
-    void drawChord(int x, int y, int width, int height, int startAngle, int spanAngle) override;
-    void drawConvexPolygon(const QPoint *points, int pointCount) override;
-    void drawEllipse(int x, int y, int width, int height) override;
-    void drawImage(const QPoint &point, const QImage &image, const QRect &source, Qt::ImageConversionFlags flags = Qt::AutoColor) override;
-    void drawLine(int x1, int y1, int x2, int y2) override;
-//    void drawPath(const QPainterPath &path) override;
-//    void fillPath(const QPainterPath &path, int colorId) override;
-//    void strokePath(const QPainterPath &path, int colorId) override;
-    void drawPie(int x, int y, int width, int height, int startAngle, int spanAngle) override;
-    void drawPoint(int x, int y) override;
-    void drawPolygon(const QPoint *points, int pointCount, Qt::FillRule fillRule = Qt::OddEvenFill) override;
-    void drawPolyline(const QPoint *points, int pointCount) override;
-    void drawRoundedRect(int x, int y, int w, int h, qreal xRadius, qreal yRadius, Qt::SizeMode mode = Qt::AbsoluteSize) override;
-    void eraseRect(int x, int y, int width, int height) override;
+    void setPen(int colorId);
+    void resetPen();
+    void setBrush(int colorId);
+    void resetBrush();
+    void setFont(int fontFamilyId, bool isBold, bool isItalic, bool isUnderline, int pxSize);
+    void setFontFamily(int fontFamilyId);
+    void setFontBold(bool isBold);
+    void setFontItalic(bool isItalic);
+    void setFontUnderline(bool isUnderline);
+    void setFontSize(int pxSize);
+    void drawText(int x, int y, QString text);
+    void drawText(int x, int y, int width, int height, int flags, QString text);
+    void drawRect(int x, int y, int width, int height);
+    void fillRect(int x, int y, int width, int height, int colorId);
+    void fillRect(int x, int y, int width, int height);
+    void drawArc(int x, int y, int width, int height, int startAngle, int spanAngle);
+    void drawChord(int x, int y, int width, int height, int startAngle, int spanAngle);
+    void drawConvexPolygon(const QPoint *points, int pointCount);
+    void drawEllipse(int x, int y, int width, int height);
+    void drawImage(const QPoint &point, const QImage &image, const QRect &source, Qt::ImageConversionFlags flags = Qt::AutoColor);
+    void drawLine(int x1, int y1, int x2, int y2);
+//    void drawPath(const QPainterPath &path);
+//    void fillPath(const QPainterPath &path, int colorId);
+//    void strokePath(const QPainterPath &path, int colorId);
+    void drawPie(int x, int y, int width, int height, int startAngle, int spanAngle);
+    void drawPoint(int x, int y);
+    void drawPolygon(const QPoint *points, int pointCount, Qt::FillRule fillRule = Qt::OddEvenFill);
+    void drawPolyline(const QPoint *points, int pointCount);
+    void drawRoundedRect(int x, int y, int w, int h, qreal xRadius, qreal yRadius, Qt::SizeMode mode = Qt::AbsoluteSize);
+    void eraseRect(int x, int y, int width, int height);
 
-    void setBackground(int colorId) override;
-    void setBackgroundMode(Qt::BGMode mode) override;
+    void setBackground(int colorId);
+    void setBackgroundMode(Qt::BGMode mode);
 
-//    void setClipPath(const QPainterPath &path, Qt::ClipOperation operation = Qt::ReplaceClip) override;
-    void setClipRect(int x, int y, int width, int height, Qt::ClipOperation operation = Qt::ReplaceClip) override;
-    void setClipping(bool enable) override;
+//    void setClipPath(const QPainterPath &path, Qt::ClipOperation operation = Qt::ReplaceClip);
+    void setClipRect(int x, int y, int width, int height, Qt::ClipOperation operation = Qt::ReplaceClip);
+    void setClipping(bool enable);
 
-    void setCompositionMode(QPainter::CompositionMode mode) override;
+    void setCompositionMode(QPainter::CompositionMode mode);
 
-    void setLayoutDirection(Qt::LayoutDirection direction) override;
+    void setLayoutDirection(Qt::LayoutDirection direction);
 
-    void setOpacity(qreal opacity) override;
-    void setRenderHint(QPainter::RenderHint hint, bool on) override;
+    void setOpacity(qreal opacity);
+    void setRenderHint(QPainter::RenderHint hint, bool on);
 
-    void setTransform(const QTransform &transform, bool combine) override;
-    void setViewTransformEnabled(bool enable) override;
+    void setTransform(const QTransform &transform, bool combine);
+    void setViewTransformEnabled(bool enable);
 
-    void setWorldTransform(const QTransform &matrix, bool combine) override;
-    void setWorldMatrixEnabled(bool enable) override;
+    void setWorldTransform(const QTransform &matrix, bool combine);
+    void setWorldMatrixEnabled(bool enable);
 
-    void shear(qreal sh, qreal sv) override;
+    void shear(qreal sh, qreal sv);
 
-    void translate(qreal dx, qreal dy) override;
+    void translate(qreal dx, qreal dy);
 
     //color management
-    int registerColor(QString colorString) override;
-    void freeColor(int colorId) override;
+    int registerColor(QString colorString);
+    void freeColor(int colorId);
 
     //font management
-    int registerFontFamily(QString fontFamilyName) override;
-    void freeFontFamily(int fontId) override;
+    int registerFontFamily(QString fontFamilyName);
+    void freeFontFamily(int fontId);
+
+    static QSize getStringSize(QChar* fontFamilyName, int fontSizePx, QString text);
 
 private:
     QPainter* m_qpainter;
